Split HTTP2 Channel out from upgrade channel
